+++ .gitignore
@@ -1,5 +1,6 @@
-node_modules
+node_modules/
 dist
 coverage        
 .env
 *.log   
+logs/
